Improve robustness (attempt) : allow working with atomic delta
What it fixes, suppose case:

t1: [ DELETE "mandataris has mandaat 1" ]
t2: [ INSERT "mandataris has mandaat 1", INSERT "foo a Bar", INSERT
"foo has baz", ... ]

In previous version, the information on t2 does not correctly
propagate to the right graph, as it does not contain sufficient
information (a type).

This because the TMP_INGEST_GRAPH got cleared every time information
was ingested and then propagated

Now we don't do this anymore, and the TMP_INGEST_GRAPH should be a
copy of the state of the consumed deltas, which should contain the
information to dispatch the triple to the right graph.

Anyway long commit message, hope it helps.