tweak on fix robustness
The fix was potentially not scalable on second thought. The tmp graph
would be growing, the copy cost will increase over time.

Since the dispatch is only based on a very limited set of
information (a type and potential link to bestuurseenheid), we don't
need the full tmp graph to be synced every time. Thus it can be cleared.